# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Career Explorer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Career {
  id: number;
  title: string;
  category: string;
  icon: string;
  description: string;
  stepCount: number;
  featured?: boolean;
}

export interface CareerStep {
  id: number;
  careerId: number;
  order: number;
  title: string;
  description: string;
  /** @nullable */
  tip?: string | null;
}

export interface CareerDetail {
  id: number;
  title: string;
  category: string;
  icon: string;
  description: string;
  stepCount: number;
  featured?: boolean;
  steps: CareerStep[];
}

export interface UserProfile {
  id: number;
  clerkId: string;
  displayName: string;
  grade: string;
  role: string;
  isBanned: boolean;
  /** @nullable */
  activeCareer?: number | null;
  createdAt: string;
}

export interface UserProfileUpdate {
  displayName?: string;
  grade?: string;
}

export interface CareerProgress {
  id: number;
  careerId: number;
  userId: number;
  completedSteps: number;
  totalSteps: number;
  completed: boolean;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface StartCareerInput {
  careerId: number;
}

export interface ProgressUpdate {
  completedSteps: number;
  completed?: boolean;
}

export interface Challenge {
  id: number;
  careerId: number;
  question: string;
  options: string[];
  difficulty: string;
}

export interface ChallengeAnswerInput {
  answer: string;
}

export interface ChallengeResult {
  correct: boolean;
  explanation: string;
  points: number;
}

export interface Stats {
  totalUsers: number;
  totalCareers: number;
  popularCareers: Career[];
}

export interface AdminUser {
  id: number;
  clerkId: string;
  displayName: string;
  grade: string;
  role: string;
  isBanned: boolean;
  email: string;
  createdAt: string;
}

export interface AdminUserList {
  users: AdminUser[];
  total: number;
  page: number;
  limit: number;
}

export interface BanUserInput {
  banned: boolean;
  reason?: string;
}

export interface SetRoleInput {
  role: string;
}

export interface AdminStats {
  totalUsers: number;
  totalCareers: number;
  activeThisWeek: number;
  popularCareers: Career[];
  recentUsers: AdminUser[];
}

export type ListCareersParams = {
category?: string;
};

export type GetMyChallengesParams = {
careerId?: number;
};

export type AdminListUsersParams = {
page?: number;
limit?: number;
};