# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Career Explorer API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUser } from './adminUser';
import type { Career } from './career';

export interface AdminStats {
  totalUsers: number;
  totalCareers: number;
  activeThisWeek: number;
  popularCareers: Career[];
  recentUsers: AdminUser[];
}