# lib/api-zod/src/generated/types/adminUserList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Career Explorer API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUser } from './adminUser';

export interface AdminUserList {
  users: AdminUser[];
  total: number;
  page: number;
  limit: number;
}