# lib/api-zod/src/generated/types/banUserInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Career Explorer API
 * OpenAPI spec version: 0.1.0
 */

export interface BanUserInput {
  banned: boolean;
  reason?: string;
}