# lib/api-zod/src/generated/types/careerDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Career Explorer API
 * OpenAPI spec version: 0.1.0
 */
import type { CareerStep } from './careerStep';

export interface CareerDetail {
  id: number;
  title: string;
  category: string;
  icon: string;
  description: string;
  stepCount: number;
  featured?: boolean;
  steps: CareerStep[];
}