# lib/api-zod/src/generated/types/careerProgress.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Career Explorer API
 * OpenAPI spec version: 0.1.0
 */

export interface CareerProgress {
  id: number;
  careerId: number;
  userId: number;
  completedSteps: number;
  totalSteps: number;
  completed: boolean;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
}