# lib/api-zod/src/generated/types/careerStep.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Career Explorer API
 * OpenAPI spec version: 0.1.0
 */

export interface CareerStep {
  id: number;
  careerId: number;
  order: number;
  title: string;
  description: string;
  /** @nullable */
  tip?: string | null;
}