# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Career Explorer API
 * OpenAPI spec version: 0.1.0
 */
import type { Career } from './career';

export interface Stats {
  totalUsers: number;
  totalCareers: number;
  popularCareers: Career[];
}