# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Career Explorer API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  clerkId: string;
  displayName: string;
  grade: string;
  role: string;
  isBanned: boolean;
  /** @nullable */
  activeCareer?: number | null;
  createdAt: string;
}